requirements.txt
openai-agents

pymongo[srv]

FastAPI

pydantic

uvicorn

python-dotenv